requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==2.0.12
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.0
gunicorn==20.1.0
idna==3.7
pytz==2024.1
requests==2.28.0
setuptools==73.0.1
sqlparse==0.5.1
urllib3==1.26.19
